Make a lagged block for multiview

make_block generates a lagged block with the appropriate max_lag and tau, while respecting lib (by inserting NANs, when trying to lag past lib regions)

make_block(block, t = NULL, max_lag = 3, tau = 1, lib = NULL,
  restrict_to_lib = TRUE)

Arguments

block
a data.frame or matrix where each column is a time series

t
the time index for the block

max_lag
the total number of lags to include for each variable. So if max_lag == 3, a variable X will appear with lags X[t], X[t - tau], X[t - 2*tau]

tau
the lag to use for time delay embedding

lib
a 2-column matrix (or 2-element vector) where each row specifies the first and last *rows* of the time series to use for attractor reconstruction

restrict_to_lib
whether to restrict the final lagged block to just the rows specified in lib (if lib exists)

Value

A data.frame with the lagged columns and a time column. If the original block had columns X, Y, Z and max_lag = 3, then the returned data.frame will have columns TIME, X, X_1, X_2, Y, Y_1, Y_2, Z, Z_1, Z_2.

Examples

data("block_3sp")
make_block(block_3sp[, c(2, 5, 8)])

#>     time         x_t       x_t_1       x_t_2          y_t        y_t_1
#> 1      1 -0.74186252          NA          NA -1.268103609           NA
#> 2      2  1.24488176 -0.74186252          NA  1.488887456 -1.268103609
#> 3      3 -1.91768516  1.24488176 -0.74186252 -0.113188052  1.488887456
#> 4      4 -0.96231763 -1.91768516  1.24488176 -1.106778594 -0.113188052
#> 5      5  1.33187510 -0.96231763 -1.91768516  2.385040811 -1.106778594
#> 6      6 -0.81708290  1.33187510 -0.96231763 -0.675346281  2.385040811
#> 7      7  0.74358596 -0.81708290  1.33187510 -0.010639353 -0.675346281
#> 8      8 -1.28017317  0.74358596 -0.81708290  1.019141615 -0.010639353
#> 9      9  0.55500418 -1.28017317  0.74358596 -0.194615118  1.019141615
#> 10    10 -0.07688614  0.55500418 -1.28017317  1.633441402 -0.194615118
#> 11    11  0.84284628 -0.07688614  0.55500418 -1.167448221  1.633441402
#> 12    12 -1.05008639  0.84284628 -0.07688614  0.217246161 -1.167448221
#> 13    13  0.99253205 -1.05008639  0.84284628 -0.672663789  0.217246161
#> 14    14 -0.31465609  0.99253205 -1.05008639 -0.296778866 -0.672663789
#> 15    15  1.46606158 -0.31465609  0.99253205 -0.510561933 -0.296778866
#> 16    16 -1.61357238  1.46606158 -0.31465609  1.465883289 -0.510561933
#> 17    17 -0.48008027 -1.61357238  1.46606158 -0.433485545  1.465883289
#> 18    18  0.99718933 -0.48008027 -1.61357238  0.118151313 -0.433485545
#> 19    19 -2.03403835  0.99718933 -0.48008027  0.507291830  0.118151313
#> 20    20 -1.56002008 -2.03403835  0.99718933 -1.226802982  0.507291830
#> 21    21  0.20177436 -1.56002008 -2.03403835  0.697495314 -1.226802982
#> 22    22  1.11325891  0.20177436 -1.56002008  0.862965977  0.697495314
#> 23    23 -1.86606858  1.11325891  0.20177436 -0.639584919  0.862965977
#> 24    24 -0.80994347 -1.86606858  1.11325891 -0.222526696 -0.639584919
#> 25    25  1.32585992 -0.80994347 -1.86606858  0.836702462 -0.222526696
#> 26    26 -1.42191395  1.32585992 -0.80994347  1.319094345  0.836702462
#> 27    27  0.13787645 -1.42191395  1.32585992 -1.363664740  1.319094345
#> 28    28  1.09538302  0.13787645 -1.42191395 -0.161693602 -1.363664740
#> 29    29 -1.05872900  1.09538302  0.13787645 -0.493608195 -0.161693602
#> 30    30  0.71181773 -1.05872900  1.09538302  0.096733727 -0.493608195
#> 31    31 -0.03144953  0.71181773 -1.05872900  1.072348624  0.096733727
#> 32    32  1.18060704 -0.03144953  0.71181773 -0.569276899  1.072348624
#> 33    33 -1.38774556  1.18060704 -0.03144953  1.736596897 -0.569276899
#> 34    34  0.18145135 -1.38774556  1.18060704 -0.873817764  1.736596897
#> 35    35  0.46899512  0.18145135 -1.38774556  2.095509948 -0.873817764
#> 36    36  0.22263716  0.46899512  0.18145135 -0.605192292  2.095509948
#> 37    37  0.67727417  0.22263716  0.46899512  0.009175989 -0.605192292
#> 38    38  0.04575542  0.67727417  0.22263716 -0.717945927  0.009175989
#> 39    39  1.08411887  0.04575542  0.67727417 -0.070537156 -0.717945927
#> 40    40 -1.16756732  1.08411887  0.04575542  0.440672258 -0.070537156
#> 41    41  0.87965869 -1.16756732  1.08411887 -1.644488587  0.440672258
#> 42    42  0.05598149  0.87965869 -1.16756732  1.564518782 -1.644488587
#> 43    43  0.90056662  0.05598149  0.87965869 -1.490914226  1.564518782
#> 44    44 -0.66943142  0.90056662  0.05598149 -0.079992629 -1.490914226
#> 45    45  1.38314427 -0.66943142  0.90056662  0.197416957 -0.079992629
#> 46    46 -1.34502123  1.38314427 -0.66943142  0.685564146  0.197416957
#> 47    47  0.24840616 -1.34502123  1.38314427 -0.321664160  0.685564146
#> 48    48  0.52903521  0.24840616 -1.34502123  0.784805545 -0.321664160
#> 49    49 -0.34943076  0.52903521  0.24840616 -0.864682834  0.784805545
#> 50    50  1.25752175 -0.34943076  0.52903521  0.777882787 -0.864682834
#> 51    51 -0.97860137  1.25752175 -0.34943076 -0.538724191  0.777882787
#> 52    52  0.97713909 -0.97860137  1.25752175 -0.309352105 -0.538724191
#> 53    53  0.13308112  0.97713909 -0.97860137  0.000553478 -0.309352105
#> 54    54  0.79084319  0.13308112  0.97713909 -0.012580698  0.000553478
#> 55    55 -1.18490130  0.79084319  0.13308112  0.563513365 -0.012580698
#> 56    56  0.67403841 -1.18490130  0.79084319  0.546839213  0.563513365
#> 57    57  0.15511822  0.67403841 -1.18490130  1.708684621  0.546839213
#> 58    58  0.71037390  0.15511822  0.67403841 -0.924342793  1.708684621
#> 59    59  0.37340513  0.71037390  0.15511822  0.229086666 -0.924342793
#> 60    60  0.60415269  0.37340513  0.71037390 -1.149584409  0.229086666
#> 61    61 -0.59967480  0.60415269  0.37340513  1.567761516 -1.149584409
#> 62    62  1.33770306 -0.59967480  0.60415269 -1.528225110  1.567761516
#> 63    63 -1.34648751  1.33770306 -0.59967480 -0.089572253 -1.528225110
#> 64    64  0.60314493 -1.34648751  1.33770306 -0.620200975 -0.089572253
#> 65    65  0.42916742  0.60314493 -1.34648751 -0.285002774 -0.620200975
#> 66    66  0.32629051  0.42916742  0.60314493 -0.930041843 -0.285002774
#> 67    67  0.64581285  0.32629051  0.42916742 -0.705679112 -0.930041843
#> 68    68  0.16259436  0.64581285  0.32629051 -1.115939859 -0.705679112
#> 69    69  0.55145187  0.16259436  0.64581285 -0.888741683 -1.115939859
#> 70    70  0.09140553  0.55145187  0.16259436  0.418740121 -0.888741683
#> 71    71  0.75911977  0.09140553  0.55145187 -0.190223442  0.418740121
#> 72    72 -0.65249895  0.75911977  0.09140553 -0.935922238 -0.190223442
#> 73    73  1.25577789 -0.65249895  0.75911977 -0.126204993 -0.935922238
#> 74    74 -1.06419718  1.25577789 -0.65249895  0.191106276 -0.126204993
#> 75    75  0.66283042 -1.06419718  1.25577789 -0.939949958  0.191106276
#> 76    76 -0.16963299  0.66283042 -1.06419718  1.525613482 -0.939949958
#> 77    77  0.90782931 -0.16963299  0.66283042 -0.705880544  1.525613482
#> 78    78 -1.79288298  0.90782931 -0.16963299  1.406556608 -0.705880544
#> 79    79 -0.61320669 -1.79288298  0.90782931 -0.380974249  1.406556608
#> 80    80  1.28062266 -0.61320669 -1.79288298  1.983322846 -0.380974249
#> 81    81 -1.92372702  1.28062266 -0.61320669 -0.352249423  1.983322846
#> 82    82 -1.14234186 -1.92372702  1.28062266  0.089891669 -0.352249423
#> 83    83  0.99847279 -1.14234186 -1.92372702  0.433500327  0.089891669
#> 84    84 -0.49268545  0.99847279 -1.14234186 -1.286678890  0.433500327
#> 85    85  1.04593430 -0.49268545  0.99847279 -0.275116858 -1.286678890
#> 86    86 -1.30878918  1.04593430 -0.49268545 -0.291274621 -0.275116858
#> 87    87  0.52806639 -1.30878918  1.04593430 -0.810957044 -0.291274621
#> 88    88  0.52502215  0.52806639 -1.30878918  1.248448195 -0.810957044
#> 89    89  0.05371973  0.52502215  0.52806639 -0.701166063  1.248448195
#> 90    90  1.11784642  0.05371973  0.52502215  0.468952116 -0.701166063
#> 91    91 -1.28280552  1.11784642  0.05371973  0.059064045  0.468952116
#> 92    92  0.68852656 -1.28280552  1.11784642 -0.766053926  0.059064045
#> 93    93  0.56646662  0.68852656 -1.28280552 -0.104379615 -0.766053926
#> 94    94  0.14631569  0.56646662  0.68852656 -1.748822004 -0.104379615
#> 95    95  0.19861192  0.14631569  0.56646662  0.405515125 -1.748822004
#> 96    96  0.22442897  0.19861192  0.14631569  0.119914373  0.405515125
#> 97    97  0.73410557  0.22442897  0.19861192 -0.941378801  0.119914373
#> 98    98  0.09486126  0.73410557  0.22442897 -0.335901057 -0.941378801
#> 99    99  0.93035993  0.09486126  0.73410557  0.287677315 -0.335901057
#> 100  100 -0.96801255  0.93035993  0.09486126  0.862975521  0.287677315
#> 101  101  0.93071059 -0.96801255  0.93035993 -0.267914232  0.862975521
#> 102  102 -1.28770657  0.93071059 -0.96801255  0.068134489 -0.267914232
#> 103  103  0.43940116 -1.28770657  0.93071059 -1.335254665  0.068134489
#> 104  104 -0.11938397  0.43940116 -1.28770657  0.667123167 -1.335254665
#> 105  105  1.45525912 -0.11938397  0.43940116 -0.767410460  0.667123167
#> 106  106 -1.45218180  1.45525912 -0.11938397  0.982509477 -0.767410460
#> 107  107 -0.21290047 -1.45218180  1.45525912 -0.736245343  0.982509477
#> 108  108  1.14032034 -0.21290047 -1.45218180  2.007073044 -0.736245343
#> 109  109 -1.15227562  1.14032034 -0.21290047 -0.456951102  2.007073044
#> 110  110  0.58577563 -1.15227562  1.14032034 -0.993882720 -0.456951102
#> 111  111  0.02267311  0.58577563 -1.15227562  0.138996414 -0.993882720
#> 112  112  0.87747088  0.02267311  0.58577563 -1.673652574  0.138996414
#> 113  113 -0.51915238  0.87747088  0.02267311 -0.197281559 -1.673652574
#> 114  114  1.13955107 -0.51915238  0.87747088 -1.187269637 -0.197281559
#> 115  115 -1.46183368  1.13955107 -0.51915238  0.542356904 -1.187269637
#> 116  116  0.34190993 -1.46183368  1.13955107 -0.656065970  0.542356904
#> 117  117  0.79146866  0.34190993 -1.46183368  2.299303509 -0.656065970
#> 118  118 -0.76943194  0.79146866  0.34190993 -1.532904554  2.299303509
#> 119  119  0.81190794 -0.76943194  0.79146866  0.476550645 -1.532904554
#> 120  120 -0.71242830  0.81190794 -0.76943194  0.947024108  0.476550645
#> 121  121  1.48320041 -0.71242830  0.81190794 -1.553976456  0.947024108
#> 122  122 -1.36275758  1.48320041 -0.71242830  0.463763502 -1.553976456
#> 123  123  0.24061552 -1.36275758  1.48320041 -2.133467193  0.463763502
#> 124  124  0.76833424  0.24061552 -1.36275758  1.928618784 -2.133467193
#> 125  125 -0.53266756  0.76833424  0.24061552 -0.538830526  1.928618784
#> 126  126  1.41743707 -0.53266756  0.76833424 -1.654048123 -0.538830526
#> 127  127 -1.26387115  1.41743707 -0.53266756  1.066558206 -1.654048123
#> 128  128  0.51690746 -1.26387115  1.41743707 -0.518649567  1.066558206
#> 129  129  0.47668533  0.51690746 -1.26387115  2.345343637 -0.518649567
#> 130  130  0.16809294  0.47668533  0.51690746 -1.519742149  2.345343637
#> 131  131  0.51582485  0.16809294  0.47668533  1.263668023 -1.519742149
#> 132  132  0.10683009  0.51582485  0.16809294 -0.631408678  1.263668023
#> 133  133  0.93927634  0.10683009  0.51582485  0.100884000 -0.631408678
#> 134  134 -0.96980667  0.93927634  0.10683009  0.253190687  0.100884000
#> 135  135  0.81690833 -0.96980667  0.93927634 -0.419619329  0.253190687
#> 136  136 -0.87054460  0.81690833 -0.96980667  0.222154272 -0.419619329
#> 137  137  1.34276297 -0.87054460  0.81690833  0.273686061  0.222154272
#> 138  138 -1.39020984  1.34276297 -0.87054460  1.153725033  0.273686061
#> 139  139  0.02815547 -1.39020984  1.34276297 -2.007422043  1.153725033
#> 140  140  1.27393780  0.02815547 -1.39020984  0.679740675 -2.007422043
#> 141  141 -0.94907719  1.27393780  0.02815547 -0.757611847  0.679740675
#> 142  142  1.09818464 -0.94907719  1.27393780  1.076221592 -0.757611847
#> 143  143 -0.47425278  1.09818464 -0.94907719 -0.433643360  1.076221592
#> 144  144  1.29601073 -0.47425278  1.09818464  0.068498026 -0.433643360
#> 145  145 -1.31714070  1.29601073 -0.47425278  0.524129750  0.068498026
#> 146  146  0.53170032 -1.31714070  1.29601073 -1.443591040  0.524129750
#> 147  147  0.49732658  0.53170032 -1.31714070  1.228474611 -1.443591040
#> 148  148  0.10153860  0.49732658  0.53170032 -1.611117765  1.228474611
#> 149  149  1.00452662  0.10153860  0.49732658 -0.209204883 -1.611117765
#> 150  150 -0.71311963  1.00452662  0.10153860 -0.359246366 -0.209204883
#> 151  151  1.18961020 -0.71311963  1.00452662 -1.509471558 -0.359246366
#> 152  152 -1.33063668  1.18961020 -0.71311963  0.171641242 -1.509471558
#> 153  153  0.53715683 -1.33063668  1.18961020 -1.987462858  0.171641242
#> 154  154  0.35739285  0.53715683 -1.33063668  1.657537070 -1.987462858
#> 155  155  0.17233427  0.35739285  0.53715683  0.560627062  1.657537070
#> 156  156  0.40022673  0.17233427  0.35739285 -0.266875720  0.560627062
#> 157  157  0.46907142  0.40022673  0.17233427  0.463695254 -0.266875720
#> 158  158  0.47399923  0.46907142  0.40022673 -0.387894069  0.463695254
#> 159  159  0.48685729  0.47399923  0.46907142 -0.516800048 -0.387894069
#> 160  160 -0.35255892  0.48685729  0.47399923 -0.799840317 -0.516800048
#> 161  161  0.74428965 -0.35255892  0.48685729  0.901866512 -0.799840317
#> 162  162 -1.13067610  0.74428965 -0.35255892 -1.043150921  0.901866512
#> 163  163  0.44382983 -1.13067610  0.74428965 -0.966408993 -1.043150921
#> 164  164  0.53229944  0.44382983 -1.13067610  1.588772570 -0.966408993
#> 165  165  0.70096937  0.53229944  0.44382983 -1.239124071  1.588772570
#> 166  166 -0.36118101  0.70096937  0.53229944  1.461516463 -1.239124071
#> 167  167  1.38659955 -0.36118101  0.70096937 -1.429021173  1.461516463
#> 168  168 -1.19127607  1.38659955 -0.36118101 -0.655800658 -1.429021173
#> 169  169  0.35761159 -1.19127607  1.38659955 -0.749296810 -0.655800658
#> 170  170  0.25941160  0.35761159 -1.19127607  0.279262086 -0.749296810
#> 171  171  0.37676786  0.25941160  0.35761159 -0.157193482  0.279262086
#> 172  172  0.25483207  0.37676786  0.25941160  0.102141104 -0.157193482
#> 173  173  0.18631702  0.25483207  0.37676786  0.641025436  0.102141104
#> 174  174  0.89411811  0.18631702  0.25483207  0.664561294  0.641025436
#> 175  175 -0.02688131  0.89411811  0.18631702 -0.354487212  0.664561294
#> 176  176  1.16305131 -0.02688131  0.89411811 -0.959938016 -0.354487212
#> 177  177 -1.27474414  1.16305131 -0.02688131  0.015739920 -0.959938016
#> 178  178  0.17991050 -1.27474414  1.16305131 -0.634825100  0.015739920
#> 179  179  0.47450700  0.17991050 -1.27474414  0.683212538 -0.634825100
#> 180  180 -0.13330921  0.47450700  0.17991050  0.006870494  0.683212538
#> 181  181  1.00060722 -0.13330921  0.47450700  0.106849168  0.006870494
#> 182  182 -0.74467390  1.00060722 -0.13330921  0.313180481  0.106849168
#> 183  183  0.99339700 -0.74467390  1.00060722 -1.376470061  0.313180481
#> 184  184 -1.32247529  0.99339700 -0.74467390  1.645157819 -1.376470061
#> 185  185  0.53577863 -1.32247529  0.99339700 -1.305362196  1.645157819
#> 186  186  0.49029211  0.53577863 -1.32247529  0.330496126 -1.305362196
#> 187  187  0.51859158  0.49029211  0.53577863 -0.136648158  0.330496126
#> 188  188  0.56420965  0.51859158  0.49029211  0.198959344 -0.136648158
#> 189  189  0.14880882  0.56420965  0.51859158 -0.618228455  0.198959344
#> 190  190  0.62693236  0.14880882  0.56420965 -0.360134070 -0.618228455
#> 191  191  0.11104523  0.62693236  0.14880882 -0.027172642 -0.360134070
#> 192  192  0.89659947  0.11104523  0.62693236 -0.915383953 -0.027172642
#> 193  193 -0.75957259  0.89659947  0.11104523  0.701270451 -0.915383953
#> 194  194  0.78442412 -0.75957259  0.89659947 -0.569933234  0.701270451
#> 195  195 -1.27741146  0.78442412 -0.75957259 -0.458297483 -0.569933234
#> 196  196  0.55873350 -1.27741146  0.78442412 -0.862807852 -0.458297483
#> 197  197  0.63397711  0.55873350 -1.27741146  0.434703441 -0.862807852
#> 198  198  0.02064414  0.63397711  0.55873350  0.631287233  0.434703441
#> 199  199  0.54892283  0.02064414  0.63397711  0.156250814  0.631287233
#> 200  200  0.04030351  0.54892283  0.02064414 -0.204528598  0.156250814
#>            y_t_2          z_t        z_t_1        z_t_2
#> 1             NA -1.863980166           NA           NA
#> 2             NA -0.481582539 -1.863980166           NA
#> 3   -1.268103609  1.535238778 -0.481582539 -1.863980166
#> 4    1.488887456 -1.492955761  1.535238778 -0.481582539
#> 5   -0.113188052 -1.119476212 -1.492955761  1.535238778
#> 6   -1.106778594  0.746657902 -1.119476212 -1.492955761
#> 7    2.385040811  0.063088675  0.746657902 -1.119476212
#> 8   -0.675346281  0.812235161  0.063088675  0.746657902
#> 9   -0.010639353 -1.731141339  0.812235161  0.063088675
#> 10   1.019141615  0.357270473 -1.731141339  0.812235161
#> 11  -0.194615118 -0.080599755  0.357270473 -1.731141339
#> 12   1.633441402  1.020736743 -0.080599755  0.357270473
#> 13  -1.167448221 -0.326734623  1.020736743 -0.080599755
#> 14   0.217246161  0.307151369 -0.326734623  1.020736743
#> 15  -0.672663789 -0.226960645  0.307151369 -0.326734623
#> 16  -0.296778866  1.523424867 -0.226960645  0.307151369
#> 17  -0.510561933 -1.540322935  1.523424867 -0.226960645
#> 18   1.465883289  0.291199560 -1.540322935  1.523424867
#> 19  -0.433485545  1.487987281  0.291199560 -1.540322935
#> 20   0.118151313 -2.443141567  1.487987281  0.291199560
#> 21   0.507291830 -1.453055089 -2.443141567  1.487987281
#> 22  -1.226802982 -0.401897498 -1.453055089 -2.443141567
#> 23   0.697495314  1.284663607 -0.401897498 -1.453055089
#> 24   0.862965977 -1.983265314  1.284663607 -0.401897498
#> 25  -0.639584919 -0.088878572 -1.983265314  1.284663607
#> 26  -0.222526696  0.375582297 -0.088878572 -1.983265314
#> 27   0.836702462 -1.568808696  0.375582297 -0.088878572
#> 28   1.319094345  0.188785080 -1.568808696  0.375582297
#> 29  -1.363664740  1.007101090  0.188785080 -1.568808696
#> 30  -0.161693602 -1.734990086  1.007101090  0.188785080
#> 31  -0.493608195  0.272353833 -1.734990086  1.007101090
#> 32   0.096733727  0.132415967  0.272353833 -1.734990086
#> 33   1.072348624  1.051856981  0.132415967  0.272353833
#> 34  -0.569276899 -1.260551265  1.051856981  0.132415967
#> 35   1.736596897  0.162404132 -1.260551265  1.051856981
#> 36  -0.873817764 -0.313141874  0.162404132 -1.260551265
#> 37   2.095509948  0.487723889 -0.313141874  0.162404132
#> 38  -0.605192292  1.248306210  0.487723889 -0.313141874
#> 39   0.009175989 -0.208522300  1.248306210  0.487723889
#> 40  -0.717945927  0.992379018 -0.208522300  1.248306210
#> 41  -0.070537156 -0.668460951  0.992379018 -0.208522300
#> 42   0.440672258  1.134456953 -0.668460951  0.992379018
#> 43  -1.644488587 -0.250573861  1.134456953 -0.668460951
#> 44   1.564518782  1.567439149 -0.250573861  1.134456953
#> 45  -1.490914226 -0.822636271  1.567439149 -0.250573861
#> 46  -0.079992629  0.844851844 -0.822636271  1.567439149
#> 47   0.197416957 -1.180108739  0.844851844 -0.822636271
#> 48   0.685564146  0.092095802 -1.180108739  0.844851844
#> 49  -0.321664160  0.571996061  0.092095802 -1.180108739
#> 50   0.784805545 -0.196853164  0.571996061  0.092095802
#> 51  -0.864682834  0.334697755 -0.196853164  0.571996061
#> 52   0.777882787 -0.661837398  0.334697755 -0.196853164
#> 53  -0.538724191  0.156713140 -0.661837398  0.334697755
#> 54  -0.309352105  0.021106833  0.156713140 -0.661837398
#> 55   0.000553478  1.021589840  0.021106833  0.156713140
#> 56  -0.012580698 -0.726263679  1.021589840  0.021106833
#> 57   0.563513365 -0.130247701 -0.726263679  1.021589840
#> 58   0.546839213 -0.547197196 -0.130247701 -0.726263679
#> 59   1.708684621  0.534906868 -0.547197196 -0.130247701
#> 60  -0.924342793  0.476250759  0.534906868 -0.547197196
#> 61   0.229086666  0.715422000  0.476250759  0.534906868
#> 62  -1.149584409 -0.009870978  0.715422000  0.476250759
#> 63   1.567761516  0.469972378 -0.009870978  0.715422000
#> 64  -1.528225110 -1.502500994  0.469972378 -0.009870978
#> 65  -0.089572253  0.843343598 -1.502500994  0.469972378
#> 66  -0.620200975  1.002675239  0.843343598 -1.502500994
#> 67  -0.285002774  0.800796457  1.002675239  0.843343598
#> 68  -0.930041843  0.900459719  0.800796457  1.002675239
#> 69  -0.705679112  1.093252848  0.900459719  0.800796457
#> 70  -1.115939859  0.468702066  1.093252848  0.900459719
#> 71  -0.888741683  0.460332592  0.468702066  1.093252848
#> 72   0.418740121  0.883427040  0.460332592  0.468702066
#> 73  -0.190223442 -0.820047555  0.883427040  0.460332592
#> 74  -0.935922238  0.783485980 -0.820047555  0.883427040
#> 75  -0.126204993 -1.111759834  0.783485980 -0.820047555
#> 76   0.191106276  1.501611380 -1.111759834  0.783485980
#> 77  -0.939949958  0.150137651  1.501611380 -1.111759834
#> 78   1.525613482  0.504806810  0.150137651  1.501611380
#> 79  -0.705880544 -1.468305915  0.504806810  0.150137651
#> 80   1.406556608 -0.783359588 -1.468305915  0.504806810
#> 81  -0.380974249  0.796960188 -0.783359588 -1.468305915
#> 82   1.983322846 -1.713377605  0.796960188 -0.783359588
#> 83  -0.352249423 -0.650358698 -1.713377605  0.796960188
#> 84   0.089891669  0.187026053 -0.650358698 -1.713377605
#> 85   0.433500327 -1.123188281  0.187026053 -0.650358698
#> 86  -1.286678890  0.153510996 -1.123188281  0.187026053
#> 87  -0.275116858 -1.154203386  0.153510996 -1.123188281
#> 88  -0.291274621  0.381242348 -1.154203386  0.153510996
#> 89  -0.810957044  0.273750925  0.381242348 -1.154203386
#> 90   1.248448195  0.076374273  0.273750925  0.381242348
#> 91  -0.701166063  1.968702843  0.076374273  0.273750925
#> 92   0.468952116 -0.765847273  1.968702843  0.076374273
#> 93   0.059064045  0.910454782 -0.765847273  1.968702843
#> 94  -0.766053926  0.510071364  0.910454782 -0.765847273
#> 95  -0.104379615  0.506670273  0.510071364  0.910454782
#> 96  -1.748822004  1.119948453  0.506670273  0.510071364
#> 97   0.405515125  0.691652789  1.119948453  0.506670273
#> 98   0.119914373  0.550543952  0.691652789  1.119948453
#> 99  -0.941378801 -0.557935180  0.550543952  0.691652789
#> 100 -0.335901057  0.358353046 -0.557935180  0.550543952
#> 101  0.287677315 -0.492011273  0.358353046 -0.557935180
#> 102  0.862975521  1.294177514 -0.492011273  0.358353046
#> 103 -0.267914232 -0.785853461  1.294177514 -0.492011273
#> 104  0.068134489  0.063762219 -0.785853461  1.294177514
#> 105 -1.335254665  0.647895194  0.063762219 -0.785853461
#> 106  0.667123167  1.604248243  0.647895194  0.063762219
#> 107 -0.767410460 -0.968658587  1.604248243  0.647895194
#> 108  0.982509477  0.015682369 -0.968658587  1.604248243
#> 109 -0.736245343  1.392153685  0.015682369 -0.968658587
#> 110  2.007073044 -1.454051433  1.392153685  0.015682369
#> 111 -0.456951102  0.399425213 -1.454051433  1.392153685
#> 112 -0.993882720  0.560635335  0.399425213 -1.454051433
#> 113  0.138996414  0.567200672  0.560635335  0.399425213
#> 114 -1.673652574 -0.868464618  0.567200672  0.560635335
#> 115 -0.197281559  1.033738604 -0.868464618  0.567200672
#> 116 -1.187269637 -1.722649032  1.033738604 -0.868464618
#> 117  0.542356904 -0.809823515 -1.722649032  1.033738604
#> 118 -0.656065970  0.333645604 -0.809823515 -1.722649032
#> 119  2.299303509 -0.685764532  0.333645604 -0.809823515
#> 120 -1.532904554  0.409882806 -0.685764532  0.333645604
#> 121  0.476550645 -0.611974569  0.409882806 -0.685764532
#> 122  0.947024108  1.635103624 -0.611974569  0.409882806
#> 123 -1.553976456 -1.395489922  1.635103624 -0.611974569
#> 124  0.463763502 -0.361018301 -1.395489922  1.635103624
#> 125 -2.133467193  1.145574128 -0.361018301 -1.395489922
#> 126  1.928618784 -0.288866235  1.145574128 -0.361018301
#> 127 -0.538830526  0.610106012 -0.288866235  1.145574128
#> 128 -1.654048123 -1.373839152  0.610106012 -0.288866235
#> 129  1.066558206  1.171060077 -1.373839152  0.610106012
#> 130 -0.518649567  0.214365735  1.171060077 -1.373839152
#> 131  2.345343637  0.284955359  0.214365735  1.171060077
#> 132 -1.519742149 -0.074050171  0.284955359  0.214365735
#> 133  1.263668023  0.876992658 -0.074050171  0.284955359
#> 134 -0.631408678  0.352305622  0.876992658 -0.074050171
#> 135  0.100884000 -0.714253297  0.352305622  0.876992658
#> 136  0.253190687  1.113742009 -0.714253297  0.352305622
#> 137 -0.419619329 -0.691322521  1.113742009 -0.714253297
#> 138  0.222154272  0.359956337 -0.691322521  1.113742009
#> 139  0.273686061 -1.576741455  0.359956337 -0.691322521
#> 140  1.153725033 -0.385605649 -1.576741455  0.359956337
#> 141 -2.007422043  1.290092728 -0.385605649 -1.576741455
#> 142  0.679740675 -0.662831157  1.290092728 -0.385605649
#> 143 -0.757611847  1.172187065 -0.662831157  1.290092728
#> 144  1.076221592 -0.251227844  1.172187065 -0.662831157
#> 145 -0.433643360  2.004927062 -0.251227844  1.172187065
#> 146  0.068498026 -1.024123594  2.004927062 -0.251227844
#> 147  0.524129750  0.909233971 -1.024123594  2.004927062
#> 148 -1.443591040  0.399913096  0.909233971 -1.024123594
#> 149  1.228474611 -0.265674664  0.399913096  0.909233971
#> 150 -1.611117765  0.227930365 -0.265674664  0.399913096
#> 151 -0.209204883 -0.840436304  0.227930365 -0.265674664
#> 152 -0.359246366  1.047505858 -0.840436304  0.227930365
#> 153 -1.509471558 -1.221479208  1.047505858 -0.840436304
#> 154  0.171641242  0.637761756 -1.221479208  1.047505858
#> 155 -1.987462858  0.390599803  0.637761756 -1.221479208
#> 156  1.657537070  0.977081309  0.390599803  0.637761756
#> 157  0.560627062 -0.011093264  0.977081309  0.390599803
#> 158 -0.266875720  0.762041508 -0.011093264  0.977081309
#> 159  0.463695254  0.937330260  0.762041508 -0.011093264
#> 160 -0.387894069  0.152309229  0.937330260  0.762041508
#> 161 -0.516800048  0.368829462  0.152309229  0.937330260
#> 162 -0.799840317  0.599504114  0.368829462  0.152309229
#> 163  0.901866512 -1.241793321  0.599504114  0.368829462
#> 164 -1.043150921  1.055181396 -1.241793321  0.599504114
#> 165 -0.966408993 -0.371044501  1.055181396 -1.241793321
#> 166  1.588772570  1.457677491 -0.371044501  1.055181396
#> 167 -1.239124071  0.077746539  1.457677491 -0.371044501
#> 168  1.461516463  0.884667255  0.077746539  1.457677491
#> 169 -1.429021173 -1.465842810  0.884667255  0.077746539
#> 170 -0.655800658  0.546040042 -1.465842810  0.884667255
#> 171 -0.749296810 -0.121429178  0.546040042 -1.465842810
#> 172  0.279262086  0.156354344 -0.121429178  0.546040042
#> 173 -0.157193482 -0.129755263  0.156354344 -0.121429178
#> 174  0.102141104 -0.661936865 -0.129755263  0.156354344
#> 175  0.641025436 -0.268272622 -0.661936865 -0.129755263
#> 176  0.664561294 -0.250569485 -0.268272622 -0.661936865
#> 177 -0.354487212  1.331953318 -0.250569485 -0.268272622
#> 178 -0.959938016 -0.841976538  1.331953318 -0.250569485
#> 179  0.015739920  0.558622387 -0.841976538  1.331953318
#> 180 -0.634825100  0.651810047  0.558622387 -0.841976538
#> 181  0.683212538 -0.129441174  0.651810047  0.558622387
#> 182  0.006870494  1.138157196 -0.129441174  0.651810047
#> 183  0.106849168 -0.329241498  1.138157196 -0.129441174
#> 184  0.313180481  1.401555311 -0.329241498  1.138157196
#> 185 -1.376470061 -1.174470236  1.401555311 -0.329241498
#> 186  1.645157819 -0.212782803 -1.174470236  1.401555311
#> 187 -1.305362196  0.724614846 -0.212782803 -1.174470236
#> 188  0.330496126  1.193687307  0.724614846 -0.212782803
#> 189 -0.136648158  1.131858525  1.193687307  0.724614846
#> 190  0.198959344  0.047550486  1.131858525  1.193687307
#> 191 -0.618228455  0.360078469  0.047550486  1.131858525
#> 192 -0.360134070  0.112107551  0.360078469  0.047550486
#> 193 -0.027172642  1.390652144  0.112107551  0.360078469
#> 194 -0.915383953 -0.218608885  1.390652144  0.112107551
#> 195  0.701270451  0.634712665 -0.218608885  1.390652144
#> 196 -0.569933234 -1.153702415  0.634712665 -0.218608885
#> 197 -0.458297483  0.358857133 -1.153702415  0.634712665
#> 198 -0.862807852  1.323193261  0.358857133 -1.153702415
#> 199  0.434703441 -0.332345924  1.323193261  0.358857133
#> 200  0.631287233  1.615404550 -0.332345924  1.323193261