Applications of empirical dynamic modeling from time series.

The rEDM package is a new implementation of EDM algorithms based on research software previously developed for internal use in the Sugihara Lab (UCSD/SIO). Contains C++ compiled objects that use time delay embedding to perform state-space reconstruction and nonlinear forecasting and an R interface to those objects using Rcpp. It supports both the simplex projection method from Sugihara & May (1990) <DOI:10.1038/344734a0> and the S-map algorithm in Sugihara (1994) <DOI:10.1098/rsta.1994.0106>. In addition, this package implements convergent cross mapping as described in Sugihara et al. (2012) <DOI:10.1126/science.1227079> and multiview embedding as described in Ye & Sugihara (2016) <DOI:10.1126/science.aag0863>.

Details

This package is divided into a set of main functions to perform various analyses, as well as helper functions that perform minor tasks, such as generate data, processing output, and wrapper functions.

Main Functions:

simplex - simplex projection for univariate forecasting

s_map - S-maps for univariate forecasting

block_lnlp - simplex or S-map forecasting with a generic reconstructed state-space

ccm - convergent cross mapping (causal inference)

multiview - multi-model approach to forecasting

tde_gp - Gaussian Processes for univariate forecasting

block_gp - Gaussian Processes with a generic reconstructed state-space

Helper Functions:

compute_stats - compute forecast skill metrics

ccm_means - aggregate output of ccm by library size (`lib_size`)

make_surrogate_data - generate surrogate time series

test_nonlinearity - test for nonlinearity using surrogate time series