(univariate) Time-Delay Embedding forecasting using Gaussian Processes

tde_gp is used in the same vein as simplex or s_map to do time series forecasting using Gaussian processes. Here, the default parameters are set so that passing a time series as the only argument will run over E = 1:10 (embedding dimension) to created a lagged block, and passing in that block and all remaining arguments into block_gp.

tde_gp(time_series, lib = c(1, NROW(time_series)), pred = lib, E = 1:10,
  tau = 1, tp = 1, phi = 0, v_e = 0, eta = 0, fit_params = TRUE,
  stats_only = TRUE, save_covariance_matrix = FALSE, silent = FALSE, ...)

Arguments

time_series
either a vector to be used as the time series, or a data.frame or matrix with at least 2 columns (in which case the first column will be used as the time index, and the second column as the time series)

lib
a 2-column matrix (or 2-element vector) where each row specifies the first and last *rows* of the time series to use for attractor reconstruction

pred
(same format as lib), but specifying the sections of the time series to forecast.

E
the embedding dimensions to use for time delay embedding

tau
the lag to use for time delay embedding

tp
the prediction horizon (how far ahead to forecast)

stats_only
specify whether to output just the forecast statistics or the raw predictions for each run

silent
prevents warning messages from being printed to the R console

...
other parameters. see 'Details'

Value

If stats_only, then a data.frame with components for the parameters and forecast statistics:

E	embedding dimension

tau	time lag

tp	prediction horizon

phi	length-scale parameter

v_e	noise-variance parameter

eta	signal-variance parameter

fit_params	whether params were fitted or not

num_pred	number of predictions

rho	correlation coefficient between observations and predictions

mae	mean absolute error

rmse	root mean square error

perc	percent correct sign

p_val	p-value that rho is significantly greater than 0 using Fisher's z-transformation

model_output	data.frame with columns for the time index, observations, mean-value for predictions, and independent variance for predictions (if stats_only == FALSE or save_covariance_matrix == TRUE)

covariance_matrix	the full covariance matrix for predictions (if save_covariance_matrix == TRUE)

E

Details

See block_gp for implementation details of the Gaussian process regression.

Examples

data("two_species_model")
ts <- two_species_model$x[1:200]
tde_gp(ts, lib = c(1, 100), pred = c(101, 200), E = 5)

#>   E tau     embedding tp       phi       v_e      eta fit_params num_pred
#> 1 5   1 1, 2, 3, 4, 5  1 0.4597587 -5.685457 6.904747       TRUE       99
#>         rho        mae       rmse perc         p_val
#> 1 0.9977515 0.01071626 0.01289583    1 6.917437e-243